+++ src/main/java/tornadofx/Component.kt
@@ -339,7 +339,17 @@ abstract class UIComponent(viewTitle: String? = "") : Component(), EventTarget {
         get() = titleProperty.get() ?: ""
         set(value) = titleProperty.set(value)
 
-    fun <T : Node> fxml(location: String? = null): ReadOnlyProperty<UIComponent, T> = object : ReadOnlyProperty<UIComponent, T> {
+    /**
+     * Load an FXML file from the specified location, or from a file with the same package and name as this UIComponent
+     * if not specified. If the FXML file specifies a controller (handy for content completion in FXML editors)
+     * set the `hasControllerAttribute` parameter to true. This ensures that the `fx:controller` attribute is ignored
+     * by the loader so that this UIComponent can still be the controller for the FXML file.
+     *
+     * Important: If you specify `hasControllerAttribute = true` when infact no `fx:controller` attribute is present,
+     * no controller will be set at all. Make sure to only specify this parameter if you actually have the `fx:controller`
+     * attribute in your FXML.
+     */
+    fun <T : Node> fxml(location: String? = null, hasControllerAttribute: Boolean = false): ReadOnlyProperty<UIComponent, T> = object : ReadOnlyProperty<UIComponent, T> {
         val value: T
 
         init {
@@ -350,7 +360,11 @@ abstract class UIComponent(viewTitle: String? = "") : Component(), EventTarget {
 
             fxmlLoader = FXMLLoader(fxml).apply {
                 resources = this@UIComponent.messages
-                setController(this@UIComponent)
+                if (hasControllerAttribute) {
+                    setControllerFactory { this@UIComponent }
+                } else {
+                    setController(this@UIComponent)
+                }
             }
 
             value = fxmlLoader!!.load()